## Refactor Project 1. Add BuildPlugin and Version Catalog 2. Refactor example module 3. Update android target sdk to 34 in order to solve the compilation error caused by `androidx.compose.ui.text:1.5.0`
