+++ build-plugin/.gitignore
@@ -1,0 +1,2 @@
+/build/
+/.gradle/
