+++ build-plugin/build.gradle.kts
@@ -1,0 +1,30 @@
+plugins {
+    `kotlin-dsl`
+    `java-gradle-plugin`
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation(gradleApi())
+    implementation(kotlin("gradle-plugin"))
+}
+
+kotlin {
+    java {
+        toolchain {
+            languageVersion.set(JavaLanguageVersion.of(17))
+        }
+    }
+}
+
+gradlePlugin {
+    plugins {
+        create("VersionPlugin") {
+            id = "com.konyaco.fluent.plugin.build"
+            implementationClass = "com.konyaco.fluent.plugin.build.BuildPlugin"
+        }
+    }
+}
