--- example/android/build.gradle.kts
@@ -1,56 +1,0 @@
-plugins {
-    id("org.jetbrains.compose")
-    id("com.android.application")
-    kotlin("android")
-}
-
-repositories {
-    google()
-    mavenCentral()
-}
-
-android {
-    namespace = "com.konyaco.fluent.example"
-    compileSdk = 33
-    defaultConfig {
-        applicationId = "com.konyaco.fluent.example"
-        minSdk = 24
-        targetSdk = 33
-        versionCode = 1
-        versionName = "1.0"
-
-        vectorDrawables {
-            useSupportLibrary = true
-        }
-    }
-    buildTypes {
-        getByName("release") {
-            isMinifyEnabled = true
-            proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
-        }
-    }
-    /*buildFeatures {
-        compose = true
-    }*/
-    packagingOptions {
-        resources {
-            excludes.add("/META-INF/{AL2.0,LGPL2.1}")
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-}
-
-kotlin {
-    jvmToolchain(11)
-}
-
-dependencies {
-    implementation(compose.foundation)
-    implementation(compose.ui)
-    implementation(compose.uiTooling)
-    implementation(project(":example:common"))
-    implementation("androidx.activity:activity-compose:1.6.1")
-}
