--- example/android/src/main/kotlin/com/konyaco/fluent/example/MainActivity.kt
@@ -1,16 +1,0 @@
-package com.konyaco.fluent.example
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            ExampleTheme {
-                App()
-            }
-        }
-    }
-}
