--- example/common/src/commonMain/kotlin/com/konyaco/fluent/example/screen/TodoScreen.kt
@@ -1,15 +1,0 @@
-package com.konyaco.fluent.example.screen
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import com.konyaco.fluent.component.Text
-
-@Composable
-fun TodoScreen() {
-    Box(Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-        Text("Not implemented yet")
-    }
-}
