+++ gallery/build.gradle.kts
@@ -1,0 +1,93 @@
+import com.konyaco.fluent.plugin.build.BuildConfig
+import org.jetbrains.compose.compose
+import org.jetbrains.compose.desktop.application.dsl.TargetFormat
+
+plugins {
+    alias(libs.plugins.kotlin.multiplatform)
+    alias(libs.plugins.compose)
+    alias(libs.plugins.android.application)
+}
+
+kotlin {
+    jvm("desktop")
+    androidTarget()
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                implementation(compose.foundation)
+                implementation(project(":fluent"))
+                implementation(project(":fluent-icons-extended"))
+                implementation(compose("org.jetbrains.compose.ui:ui-util"))
+            }
+        }
+        val commonTest by getting {
+            dependencies {
+                implementation(kotlin("test"))
+            }
+        }
+        val androidMain by getting {
+            dependencies {
+                implementation(libs.androidx.activity)
+            }
+        }
+        val androidUnitTest by getting
+        val androidInstrumentedTest by getting {
+            dependencies {
+                implementation(libs.androidx.test.junit)
+            }
+        }
+        val desktopMain by getting {
+            dependencies {
+                implementation(compose.preview)
+                implementation(libs.mayakapps.window.styler)
+            }
+        }
+        val desktopTest by getting
+    }
+    jvmToolchain(BuildConfig.Jvm.jvmToolchainVersion)
+}
+
+android {
+    compileSdk = BuildConfig.Android.compileSdkVersion
+    namespace = BuildConfig.packageName + ".gallery"
+
+    defaultConfig {
+        minSdk = BuildConfig.Android.minSdkVersion
+        targetSdk = BuildConfig.Android.compileSdkVersion
+        versionCode = 1
+        versionName = "1.0"
+        vectorDrawables {
+            useSupportLibrary = true
+        }
+
+    }
+
+    buildTypes {
+        release {
+            isMinifyEnabled = true
+            proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.android.pro")
+        }
+    }
+
+    packagingOptions {
+        resources {
+            excludes.add("/META-INF/{AL2.0,LGPL2.1}")
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility = BuildConfig.Jvm.javaVersion
+        targetCompatibility = BuildConfig.Jvm.javaVersion
+    }
+}
+
+compose.desktop {
+    application {
+        mainClass = "${BuildConfig.packageName}.gallery.MainKt"
+        nativeDistributions {
+            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
+            packageName = "Compose Fluent Design Gallery"
+            packageVersion = "1.0.0"
+        }
+    }
+}
