+++ gallery/proguard-rules.android.pro
