+++ gallery/src/androidMain/kotlin/com/konyaco/fluent/gallery/MainActivity.kt
@@ -1,0 +1,16 @@
+package com.konyaco.fluent.gallery
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            GalleryTheme {
+                App()
+            }
+        }
+    }
+}
