+++ gallery/src/androidMain/res/mipmap-xxxhdpi/ic_launcher.webp
