+++ gallery/src/androidMain/res/values/colors.xml
