+++ gallery/src/androidMain/res/values/themes.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="Theme.ComposeFluentDesignGallery" parent="android:Theme.Material.Light.NoActionBar" />
+</resources>
