+++ gallery/src/androidMain/res/xml/data_extraction_rules.xml
