+++ gallery/src/androidUnitTest/kotlin/com/konyaco/fluent/gallery/GalleryUnitTest.kt
@@ -1,0 +1,17 @@
+package com.konyaco.fluent.gallery
+
+import org.junit.Test
+
+import org.junit.Assert.*
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class GalleryUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
