+++ gallery/src/commonMain/kotlin/com/konyaco/fluent/gallery/screen/TodoScreen.kt
@@ -1,0 +1,15 @@
+package com.konyaco.fluent.gallery.screen
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import com.konyaco.fluent.component.Text
+
+@Composable
+fun TodoScreen() {
+    Box(Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+        Text("Not implemented yet")
+    }
+}
